fix: fixes for UI and functional issues [SPRW-1187]
