+++ packages/@sparrow-workspaces/src/features/ai-request-explorer/components/ai-config-selection/AiConfigurations.svelte
@@ -4,7 +4,7 @@
     modelsConfigType,
   } from "@sparrow/common/types/workspace/ai-request-base";
   import type { AiModelProviderEnum } from "@sparrow/common/types/workspace/ai-request-base";
-  import { configFormat } from "@sparrow/common/types/workspace/ai-request-dto";
+  import { configFormat } from "../../constants";
   import { Button, Toggle } from "@sparrow/library/ui";
 
   export let isActive = true;
@@ -23,17 +23,12 @@
     onUpdateAiConfigurations(currSelectedModel, updatedConfig); // Call the update function to persist changes
   };
 
-  // Reset to default values
-  // const handleResetToDefault = () => {
-  //   const defaultConfig = {};
-  //   Object.entries(currentModelConfig).forEach(
-  //     ([key, metadata]: [string, any]) => {
-  //       defaultConfig[key] = metadata.defaultValue;
-  //     },
-  //   );
-  //   console.log("in reset ;>> ");
-  //   onUpdateAiConfigurations(currSelectedModel, defaultConfig);
-  // };
+  // Check if any configuration has been changed from default
+  $: hasChanges = configEntries.some(([key, metadata]) => {
+    const currentValue =
+      config[key] !== undefined ? config[key] : metadata.defaultValue;
+    return currentValue !== metadata.defaultValue;
+  });
 
   // Reset to default values
   const handleResetToDefault = () => {
@@ -94,76 +89,72 @@
 <div class="ai-config-container">
   <div
     class="d-flex justify-content-between align-items-start mb-3"
-    style="border-bottom: 0.2px solid var(--bg-ds-surface-300); height: 34px;"
+    style="border-bottom: 0.2px solid var(--bg-ds-surface-300); height: 34px; position: sticky;
+      top: 0;
+      z-index: 10;
+      padding-bottom: 5px;"
   >
     <p
       class="config-header text-ds-font-size-14 text-ds-font-weight-medium mb-0"
     >
       Configurations
     </p>
-    <!-- <button
-      class="btn btn-sm btn-link text-ds-font-size-12"
-      class:disabled={!isActive}
-      style="color: var(--bg-ds-neutral-500); text-decoration: none; font-size: 12px;"
-      on:click={handleResetToDefault}
-      disabled={!isActive}
-    >
-      Reset to default
-    </button> -->
-
     <Button
       title={"Reset to default"}
       size={"extra-small"}
       type={"teritiary-regular"}
-      disable={true}
-      onClick={() => {}}
+      disable={!isActive || !hasChanges}
+      onClick={handleResetToDefault}
     ></Button>
   </div>
 
-  {#if configEntries.length === 0}
-    <div class="text-center py-4">
-      <p class="text-muted">
-        No configuration options available for the selected model.
-      </p>
-    </div>
-  {:else}
-    {#each configEntries as [key, metadata]}
-      <div
-        class="config-item text-ds-font-size-12 mb-3"
-        class:option-disabled={!isActive}
-      >
-        <div class="d-flex justify-content-between align-items-start mb-2">
-          <div>
-            <div class="fw-medium">{metadata.displayName || key}</div>
-            <div class="config-desc text-ds-font-size-11">
-              {metadata.description || ""}
+  <div class="config-section">
+    {#if configEntries.length === 0}
+      <div class="text-center py-4">
+        <p class="text-muted">
+          No configuration options available for the selected model.
+        </p>
+      </div>
+    {:else}
+      {#each configEntries as [key, metadata]}
+        <div
+          class="config-item text-ds-font-size-12 mb-3"
+          class:option-disabled={!isActive}
+        >
+          <div
+            class="d-flex justify-content-between align-items-start mb-2 gap-1"
+          >
+            <div>
+              <div class="fw-medium">{metadata.displayName || key}</div>
+              <div class="config-desc text-ds-font-size-11">
+                {metadata.description || ""}
+              </div>
             </div>
-          </div>
 
-          {#if metadata.type === "boolean"}
-            <Toggle
-              id={`toggle-field-${metadata.displayName}`}
-              isActive={getCurrentValue(key, metadata)}
-              disabled={!isActive}
-              onChange={(event) =>
-                handleToggleChange(key, event.target.checked)}
-            />
-          {:else if metadata.type === "number"}
-            <div class="config-value d-flex justify-content-end">
-              <input
-                id={`config-field-${metadata.displayName}`}
-                type="number"
-                class="form-control form-control-sm config-input"
-                value={getCurrentValue(key, metadata) || 0}
-                min={metadata.min}
-                max={metadata.max}
-                step={key === "maxTokens" ? 1 : 0.1}
+            {#if metadata.type === "boolean"}
+              <Toggle
+                id={`toggle-field-${metadata.displayName}`}
+                isActive={getCurrentValue(key, metadata)}
                 disabled={!isActive}
-                on:change={(e) => handleInputChange(key, e)}
+                onChange={(event) =>
+                  handleToggleChange(key, event.target.checked)}
               />
-            </div>
-          {:else}
-            <!-- <div class="config-value d-flex justify-content-end">
+            {:else if metadata.type === "number"}
+              <div class="config-value d-flex justify-content-end">
+                <input
+                  id={`config-field-${metadata.displayName}`}
+                  type="number"
+                  class="form-control form-control-sm config-input"
+                  value={getCurrentValue(key, metadata) || 0}
+                  min={metadata.min}
+                  max={metadata.max}
+                  step={key === "maxTokens" ? 1 : 0.1}
+                  disabled={!isActive}
+                  on:change={(e) => handleInputChange(key, e)}
+                />
+              </div>
+            {:else}
+              <!-- <div class="config-value d-flex justify-content-end">
               <input
                 type="text"
                 class="form-control form-control-sm config-input"
@@ -172,21 +163,26 @@
                 on:input={(e) => handleInputChange(key, e)}
               />
             </div> -->
-          {/if}
+            {/if}
+          </div>
         </div>
-      </div>
-    {/each}
-  {/if}
+      {/each}
+    {/if}
+  </div>
 </div>
 
 <style>
   .ai-config-container {
     font-family: inter, "sans-serif";
-  }
-  .config-header {
+    height: 100%; /* Make sure container has defined height */
+    display: flex;
+    flex-direction: column;
   }
 
-  .config-item {
+  .config-section {
+    flex: 1; /* Take remaining space */
+    overflow-y: auto; /* Enable vertical scrolling */
+    padding-right: 4px; /* Add some padding for scrollbar */
   }
 
   .config-desc {
@@ -211,7 +207,7 @@
     color: var(--white-color);
     border: 1px solid var(--bg-ds-surface-300);
     border-radius: 4px;
-    padding: 4px 8px;
+    padding: 4px 0px;
   }
 
   .config-input:focus {
