+++ packages/@sparrow-workspaces/src/features/ai-request-explorer/components/model-options/ModelOptions.svelte
@@ -76,7 +76,7 @@
       icon: GoogleVectorIcon,
       models: [
         { name: "Gemini 1.5 Flash", id: "gemini-1.5-flash" },
-        { name: "Gemini 1.5 Flash 8b", id: "gemini-1.5-flash-8b" },
+        { name: "Gemini 1.5 Flash 8B", id: "gemini-1.5-flash-8b" },
         { name: "Gemini 1.5 Pro", id: "gemini-1.5-pro" },
         { name: "Gemini 2.0 Flash", id: "gemini-2.0-flash" },
       ],
