+++ packages/@sparrow-workspaces/src/features/ai-request-explorer/constants/index.ts
@@ -1,0 +1,1 @@
+export { configFormat, disabledModelFeatures } from "./AiModelConfigurations"
